User: refactor live user component (#3196)
* refactor live user component

* move elements and fix types
